src/requirements.txt
paho_mqtt==2.1.0
pyModbusTCP==0.3.0
